% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compare_models}
\alias{compare_models}
\title{Compare Models}
\usage{
compare_models(model1, model2)
}
\arguments{
\item{model1}{An object of class "rsm_model".}

\item{model2}{An object of class "rsm_model".}
}
\value{
A tibble with comparison results including F-statistic and p-value.
}
\description{
Compares two RSM models using ANOVA F-test.
}
