% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{fit_first_order_interaction}
\alias{fit_first_order_interaction}
\title{Fit First-Order Model with Interaction}
\usage{
fit_first_order_interaction(data, response, factors)
}
\arguments{
\item{data}{A data frame or tibble containing the response and factor variables.}

\item{response}{Character string with the name of the response variable.}

\item{factors}{Character vector with names of factor variables (must be in coded scale).}
}
\value{
An object of class "rsm_model" containing model results.
}
\description{
Fits a first-order model with all two-way interaction terms.
}
\examples{
data <- tibble::tibble(
  Y = c(6.66, 6.30, 6.32, 5.92, 6.09, 6.22, 5.29, 6.67, 5.67),
  P_coded = c(-1, -1, 1, 1, 0, -1.68, 1.68, 0, 0),
  S_coded = c(-1, 1, -1, 1, 0, 0, 0, -1.68, 1.68)
)

model <- fit_first_order_interaction(data, response = "Y", factors = c("P_coded", "S_coded"))

}
