% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{fit_second_order}
\alias{fit_second_order}
\title{Fit Second-Order (Quadratic) Response Surface Model}
\usage{
fit_second_order(data, response, factors)
}
\arguments{
\item{data}{A data frame or tibble containing the response and factor variables.}

\item{response}{Character string with the name of the response variable.}

\item{factors}{Character vector with names of factor variables (must be in coded scale).}
}
\value{
An object of class "rsm_model" containing model results.
}
\description{
Fits a complete second-order polynomial model including linear, quadratic,
and interaction terms.
}
\details{
The model fitted is:

\deqn{Y = \beta_0 + \sum \beta_i X_i + \sum \beta_{ii} X_i^2 + \sum_{i<j} \beta_{ij} X_i X_j + \epsilon}
}
\examples{
data <- tibble::tibble(
  Y = c(6.66, 6.30, 6.32, 5.92, 6.09, 6.22, 5.29, 6.67, 5.67),
  P_coded = c(-1, -1, 1, 1, 0, -1.68, 1.68, 0, 0),
  S_coded = c(-1, 1, -1, 1, 0, 0, 0, -1.68, 1.68)
)

model <- fit_second_order(data, response = "Y", factors = c("P_coded", "S_coded"))

}
