% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_design_matrix}
\alias{get_design_matrix}
\title{Extract Model Matrix}
\usage{
get_design_matrix(rsm_model)
}
\arguments{
\item{rsm_model}{An object of class "rsm_model".}
}
\value{
A matrix with the design matrix used in the model.
}
\description{
Extracts the design matrix (X matrix) from an RSM model.
}
