% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_fitted_values}
\alias{get_fitted_values}
\title{Extract Fitted Values}
\usage{
get_fitted_values(rsm_model)
}
\arguments{
\item{rsm_model}{An object of class "rsm_model".}
}
\value{
A numeric vector with fitted values.
}
\description{
Extracts fitted values from a fitted RSM model.
}
