% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_residuals}
\alias{get_residuals}
\title{Extract Residuals}
\usage{
get_residuals(rsm_model)
}
\arguments{
\item{rsm_model}{An object of class "rsm_model".}
}
\value{
A numeric vector with residuals.
}
\description{
Extracts residuals from a fitted RSM model.
}
