% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otimizacao_corrigida.R
\name{otimizacao_insumos}
\alias{otimizacao_insumos}
\title{Otimização de Uso de Insumos (DMET e DMEE)}
\usage{
otimizacao_insumos(
  modelo_quadratico,
  preco_insumo,
  preco_produto,
  verbose = TRUE
)
}
\arguments{
\item{modelo_quadratico}{Objeto de resultado da função \code{ajustar_quadratico}.}

\item{preco_insumo}{Preço unitário do insumo (dose).}

\item{preco_produto}{Preço unitário do produto (resposta).}

\item{verbose}{Lógico. Gera prints detalhados (padrão: TRUE).}
}
\value{
Lista contendo um data frame com os resultados de otimização e uma lista de gráficos.
}
\description{
Calcula a Dose de Máxima Eficiência Técnica (DMET) e a Dose de Máxima Eficiência
Econômica (DMEE) para modelos de regressão quadrática (Y = b0 + b1\emph{X + b2}X^2).
}
