% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_diagnostics}
\alias{plot_diagnostics}
\title{Model Diagnostics}
\usage{
plot_diagnostics(rsm_model)
}
\arguments{
\item{rsm_model}{An object of class "rsm_model".}
}
\value{
A list of ggplot2 objects with diagnostic plots:
\itemize{
\item Residuals vs Fitted
\item Q-Q plot
\item Scale-Location plot
\item Residuals vs Leverage
}
}
\description{
Produces diagnostic plots for model validation.
}
