% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_isoquants}
\alias{plot_isoquants}
\title{Plot Isoquants (Contour Lines)}
\usage{
plot_isoquants(
  rsm_model,
  factor1 = NULL,
  factor2 = NULL,
  fixed_values = NULL,
  n_grid = 100,
  n_levels = 10
)
}
\arguments{
\item{rsm_model}{An object of class "rsm_model".}

\item{factor1}{Character name of first factor.}

\item{factor2}{Character name of second factor.}

\item{fixed_values}{Named list with values for other factors.}

\item{n_grid}{Integer number of grid points. Default is 100.}

\item{n_levels}{Integer number of contour levels. Default is 10.}
}
\value{
A ggplot2 object with isoquants.
}
\description{
Creates a detailed contour plot showing isoquants (lines of constant response)
for two factors.
}
