% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_steepest_path}
\alias{plot_steepest_path}
\title{Plot Steepest Path}
\usage{
plot_steepest_path(
  steepest_path_obj,
  rsm_model,
  factor1 = NULL,
  factor2 = NULL
)
}
\arguments{
\item{steepest_path_obj}{An object of class "steepest_path".}

\item{rsm_model}{The original RSM model (for background contours).}

\item{factor1}{Character name of first factor.}

\item{factor2}{Character name of second factor.}
}
\value{
A ggplot2 object with contours and the steepest path.
}
\description{
Visualizes the path of steepest ascent/descent on a contour plot.
}
