% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical_analysis.R
\name{print.canonical_analysis}
\alias{print.canonical_analysis}
\title{Print Method for Canonical Analysis}
\usage{
\method{print}{canonical_analysis}(x, ...)
}
\arguments{
\item{x}{An object of class "canonical_analysis".}

\item{...}{Additional arguments (unused).}
}
\description{
Print Method for Canonical Analysis
}
