% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{print.optimization_result}
\alias{print.optimization_result}
\title{Print Method for Optimization Results}
\usage{
\method{print}{optimization_result}(x, ...)
}
\arguments{
\item{x}{An object of class "optimization_result".}

\item{...}{Additional arguments (unused).}
}
\description{
Print Method for Optimization Results
}
