% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{print.rsm_model}
\alias{print.rsm_model}
\title{Print Method for RSM Models}
\usage{
\method{print}{rsm_model}(x, ...)
}
\arguments{
\item{x}{An object of class "rsm_model".}

\item{...}{Additional arguments (unused).}
}
\description{
Print Method for RSM Models
}
