% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steepest_path.R
\name{print.steepest_path}
\alias{print.steepest_path}
\title{Print Method for Steepest Path}
\usage{
\method{print}{steepest_path}(x, ...)
}
\arguments{
\item{x}{An object of class "steepest_path".}

\item{...}{Additional arguments (unused).}
}
\description{
Print Method for Steepest Path
}
