\name{soy_fertility}
\docType{data}
\alias{soy_fertility}
\title{Soybean fertility response surface dataset}
\description{
Experimental data used to demonstrate response surface methodology
applied to soybean fertility experiments.
}
\usage{
soy_fertility
}
\format{
A data frame with the following columns:
\describe{
  \item{N}{Nitrogen dose}
  \item{P}{Phosphorus dose}
  \item{K}{Potassium dose}
  \item{yield}{Soybean yield}
}
}
\source{
Simulated or experimental data for teaching response surface methodology.
}
