% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_analysis.R
\name{summary.anova_rsm}
\alias{summary.anova_rsm}
\title{Summary Method for ANOVA RSM Results}
\usage{
\method{summary}{anova_rsm}(object, ...)
}
\arguments{
\item{object}{An object of class "anova_rsm".}

\item{...}{Additional arguments (unused).}
}
\description{
Summary Method for ANOVA RSM Results
}
