% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical_analysis.R
\name{summary.canonical_analysis}
\alias{summary.canonical_analysis}
\title{Summary Method for Canonical Analysis}
\usage{
\method{summary}{canonical_analysis}(object, ...)
}
\arguments{
\item{object}{An object of class "canonical_analysis".}

\item{...}{Additional arguments (unused).}
}
\description{
Summary Method for Canonical Analysis
}
