% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{summary.rsm_model}
\alias{summary.rsm_model}
\title{Summary Method for RSM Models}
\usage{
\method{summary}{rsm_model}(object, ...)
}
\arguments{
\item{object}{An object of class "rsm_model".}

\item{...}{Additional arguments (unused).}
}
\description{
Summary Method for RSM Models
}
