% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steepest_path.R
\name{summary.steepest_path}
\alias{summary.steepest_path}
\title{Summary Method for Steepest Path}
\usage{
\method{summary}{steepest_path}(object, ...)
}
\arguments{
\item{object}{An object of class "steepest_path".}

\item{...}{Additional arguments (unused).}
}
\description{
Summary Method for Steepest Path
}
